pyproject.toml
[tool.poetry]
packages = [
    { include = "gendiff" },
]

name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["valek <valekgodovan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
